requirements.txt
psycopg2-binary
pgvector
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai
Django>=3.2,<4.0
django-composer